requirements.txt
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
db-dtypes==1.2.0
eralchemy2==1.3.8
fonttools==4.49.0
google-api-core==2.17.1
google-auth==2.28.1
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
graphviz==0.20.3
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pillow==10.2.0
protobuf==4.25.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pygraphviz==1.12
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
rsa==4.9
six==1.16.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
flask==3.0.3
waitress==3.0.0